Restructuring AttributeEngine directories
