+++ src/AttributeEngine/Attributes/AttributeInterface.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace OpenDialogAi\AttributeEngine;
+namespace OpenDialogAi\AttributeEngine\Attributes;
 
 /**
  * An Attribute is a perceivable feature of the environment (username, time, etc)
