+++ src/AttributeEngine/Attributes/BasicAttribute.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace OpenDialogAi\AttributeEngine;
+namespace OpenDialogAi\AttributeEngine\Attributes;
 
 /**
  * BasicAttribute is a simple implementation of the AttributeInterface that
